requirements.txt
altair==5.0.1
attrs==24.2.0
backports.zoneinfo==0.2.1
bidict==0.22.1
blinker==1.6.3
brevo-python==1.1.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cycler==0.11.0
dash==2.15.0
dash-ag-grid==31.3.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
dominate==2.9.1
et-xmlfile==1.1.0
eventlet==0.38.2
Flask==2.2.5
Flask-Bootstrap==3.3.7.1
fonttools==4.38.0
fpdf2==2.8.3
gitdb==4.0.12
gitpython==3.1.45
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.3.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jinja2==3.1.6
jsonschema==4.17.3
kiwisolver==1.4.5
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.5.5.64
openpyxl==3.1.3
packaging==23.2
pandas==1.3.5
pathlib==1.0.1
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
plotly==5.18.0
proto-plus==1.26.1
protobuf==4.24.4
psutil==7.0.0
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b1
pygments==2.17.2
PyJWT==2.8.0
Pympler==1.1
pyparsing==3.1.4
pypdf==5.0.0
pypdf2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.11.0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.8.1
rsa==4.9.1
seaborn==0.12.2
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
streamlit==1.23.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.2
typing-extensions==4.7.1
tzdata==2025.2
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==2.0.7
validators==0.20.0
visitor==0.1.3
watchdog==3.0.0
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
